Ai_model/clipping_ssbd_videos/autism_gesture_api/requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python-headless
pillow
numpy